Saving only the stash number to the data file
It complicates the script a little but it simplifies the file which is
preferable. If popped stash is just a commit and not a real stash, it's
not saved at all and the file is empty. (This will change soon anyway
when pop and apply will be split into separate commands.)