Made test scripts to handle the new file structure
The most changes were done to the `tests/run.sh` which now shows each
category (directory) of tests separately.
The tests run and pass now BUT this doesn't mean the commands work like
intended. They are still copies of the old `git istash` instead of being
the proper `git istash-apply` and `git istash-pop`. (And the third file
is still not only unchanged but also unused.)