Made it clear that I have no idea what I'm writing
OK, so the thing is that I didn't want to name the command "istash".
When it was only for my private use, I was calling it "siupa stash" or
"sstash" but this was too self-indulgent so I've changed it before
publishing the script on GitHub. I came with a new name "improved stash"
but idn... I don't care about the name that much. It is the official
name but you can treat it as a bad backronym for "istash" and come up
with your own like I just did.