Split `commons.sh` into 7 smaller files
It was getting too big to navigate.
Also, assertions for test files were removed from tests for `git istash
push` because they are relevant only for `apply` and `push` and not all
assertions get sourced in every test now. Hopefully it will make the
execution of the tests a little faster?