Added tests for stashing with (almost) no changes
The script should not create a stash when there are no changes but it
should return 0 (because `git stash` does that). (This may be changed
later.)
When there are no changes but they are untracked files, stashing with
`--include-untracked` should work as if there were changes.

One of the test is partially disabled because of a bug in the official
command `git stash push`. The test will be reenabled when either the bug
is fixed or this script stops using that Git command internally (which
is the plan anyway).