Parametrized the HEAD type in all tests
This mean that there are no separate tests for stashing on a proper
branch, a detached HEAD and an orphan branch. All three are merged into
a single test, which removed a lot of duplicated code (over 12000 lines)
without making the tests less through. (If anything, they are more so
now because in order to merge these 3 approaches, the tests were
normalized (some redundant parts removed) which in some cases resulted
in creation of new tests to still cover all the cases that were covered
before, but this time for all three types of HEAD.)

Also, some internal code of assertions were modified to be consistent
between different functions. (Most notably, now the orphan branch is
always indicated by prefixing a branch name with `~`.)