Added sections to all the tests
They make tests more understandable if used correctly but primarily they
are displayed in the failure messages, which makes it easier to find the
line that is failing.

OK, story time:
When you have 5 assertions in a single test, checking for existence of
the same file on different stages and you get the message that the file
is missing, it's really annoying to add echos everywhere just to find
which of the assertions was failing.
It would be very nice to just see the line number in the failure report.
Well, it turns out that this simple thing is not really achievable in a
POSIX script; at least not in an easy way. So obviously, I didn't give
up and I tried the hard way instead. Ignoring distant echoes of Tony The
Pony's hoof beats, I've started to write regexes that would parse each
tests and add an instruction that saves the current line number, in each
place where it wouldn't break anything... and I almost succeeded. (I had
to create a complex state machine but it was almost working.) I think I
would be able to iron out the kinks but... it needed around 10 seconds
to convert a single test. (I'm using a very low-end machine for testing,
but still.) Obviously that was a little too much for using it in tests
(that should run as fast as possible to not discourage people from using
them). And even if I tried to do something clever, like running it only
for failing tests, it would still be 200 lines of complex code just to
have something that would probably keep breaking when new tests were
added.
At this point I had to come to my senses and I went for the second best
solution that I could think of, which is naming section of tests. Given
that the same assertion is unlikely to be used in the same section of a
test, they would allow to reliably find the failing assertions, although
of course, naming sections has some the same problems that commenting
the code tends to have - namely, they tend to rot with time and lie.
Still, it's better than nothing. It just requires to be more careful
while editing the tests.
While I didn't end up using the code, I will commit it here for anyone
who's curious in the file `add-line-numbers.sh`. (It will be deleted
next commit, to not needlessly litter the repository.)