Hiding individual parametric runs by default
Individual parametric runs are not shown by default now if they pass.
Only the summary of the entire test is displayed, which solves the
problem of too long output.
A new option `--verbose` was added that reverts the script to the
previous behavior. (It's cannot be used together with `--quiet`.)