Made `istash push` handle weird character in paths
Now the paths can contain whitespaces (including line breaks) and quite
possibly all other weird characters (to be verified).

This implementation uses `hexdump`, which is not part of POSIX but it
seems to be more likely to be present than `od`. The final version will
check for the presence of `od` before fallbacking to `hexdump`.