Changed how the commits for stash are created
There are additional commits now, for changes that are skipped in the
created stash. Commits are now rearranged into 2 (mostly) independent
trees, one containing changes for stash and the other containing changes
that stay in the working directory.
This approach doesn't make much difference for currently passing tests
but it will be essential for `--patch` and pathspecs.