Fixed some potentially incorrent calls of `printf`
They were probably fine but a variable was passed as the first argument
so potentially the first character could be "-", which would make it be
confused with an option. "--" was added in those cases just to be sure.
There were also one case of `printf` that could get no arguments at all.