Added some delay to tests to fix Windows failures
This is not a pretty solution but this problem doesn't seem to be
solvable. The shell just eats all the input if it has a chance so
the input needs to be served in parts to avoid it.
It shouldn't cause problems too often since the part that breaks on that
is supposed to be interactive but you know someone will eventually use
this in a script. (Probably me.)