Added options `--staged` and `--leave-staged`
`--staged` works (mostly) the same as the same option from `git stash`,
although, it's most specific in how it works - it excludes tracked but
non-staged files from the stash creation.
`--leave-staged` is a counterpart of that, that excludes only staged
files. (Similar to keep index but the index isn't stashed at all.)