Added `--literal-pathspecs` to a few lines of code
Some of these places are a real bugs and in some places that doesn't
really change anything but the option was just added on principle.
A new tests were added but they pass in the previous version too because
there is currently a bug in how Git parses pathspecs. It was reported
and it will be fixed in a future release.