Adjusted number of parametric test runs
It's still a little overkill right now but the number of runs for the
worst cases of tests is now reduced. Still, the limit had to increase in
because a few tests reach almost 10000 runs on highest meticulousnesses.
Beside modifying specific tests, meticulousness of level 5 works now a
little differently - if there are alternative spelling of some options
defined (like `--leave` instead of `--leave-staged`), only those are
used instead of those *in addition to* the normal spellings.