Replaced `git reset --hard` in tests
These lines are replaces with a new function `remove_all_changes` which
calls the reset command and also `git clean -dfx`.
This shortens the tests a little because the list of expected files at
the end is smaller. The main reason for doing it, however, is adding
more `push`-like commands in the next commit. They will leave different
files in the working copy, so it's the most convenient to remove all
the leftover in order to use the same tests. (And these leftover files
aren't useful for tests. They were left before because I didn't have
any strong opinion on them.)