Added sub-commands `create`, `snatch` and `save`
Those are commands similar to `push` but their differs in whether the
changes are removed from the working directory and whether the stash is
stored (using `git stash store`):
- create - doesn't remove changes nor stores the stash,
- snatch - removes changes and doesn't store the stash,
- save - doesn't remove changes and stores the stash,
- push - removes changes and stores the stash.

The tests in `tests/creating` was extended to test those sub-commands
too, which unfortunately increases their runtime 4 times.
Reducing this time (if possible) is the next goal.