Reduced the amount of parametric runs of tests
The previous approach was to run each test with every set of options
that was compatible with it. Sometimes some specific operation doesn't
work correctly only in a very particular case and testing every case I
could think of increases the probability of catching them. (It's what
they call "going completely overboard with tests" but there were a few
actual situations when only a few parametric runs of a specific test
didn't pass, and I even caught a few bugs in Git itself.)
However, after basically quadrupling the test time by adding the 3 new
sub-commands, even I had to admit that this may be a little too much.

I went through all the tests that took over 200 seconds to run (which is
roughly 200 parametric runs on my machine, depending on test complexity)
and I chose a fixed value for every parameter that was not relevant to
the test. (I.e. that I didn't expect this test to run the code behind
this option in any significant capacity, so it was unlikely to find any
bugs in it.) I tried to do it in a way that will still check a wide
variety of option combinations across all the tests.