Fixed verbosity levels in `tests/run.sh`
There were some code that was breaking options `--quiet` and `--quieter`
in the test script. This is may be a little Chesterton's Fence type of
situation but I think it was mostly related to the special case when the
test result was supposed to be displayed a little different where the
test wasn't parametrised and it failed. I just removed that special case
because pretty much all tests are parametrized now. (It looks a little
weird now without parameters, though.)