"Optimized" tests by using some code magix
*OK, this is not a serious commit.* (The next commit reverts it.
It's committed only as a memento.)

Basically, the problem is that each of the almost 300 tests has to load
all the helper functions and assertions. The assumption was that this
takes a non-negligible amount of time. The solution was to load all the
function only one time, in the main test script, before running any
tests. Then the calls to tests needed to be changed from something like
`sh test.sh` to something like `(. test.sh)`, because POSIX shell does
not have any way to export functions to other script called normally.

This clever optimization allowed to decrease the runtime to 114%.