Moved some tests to the category "main_script"
The tests were in the category "push" but they really check if the main
script can infer `push` when it's not explicitly stated.

Also, removed a test that was obsolete for a long time.
(It checked if the command will exit with error of the is no subcommand,
which is not the correct behavior anymore. The tests happened to pass
only because there were a different error - no files to stash.)