Loading all utilities and assertions for each test
Some of the files were loaded only in certain tests categories to avoid
overhead of parsing the files every time. However, it seems that this
doesn't actually helps.
It will be easier now to use those tools in some special case tests that
are in categories, that don't use them normally.