`assert_exit_code` in tests also captures outputs
This makes the function do a little more than its name would suggest,
but it will eliminate a huge amount of boiler plate. The values of the
output streams will be needed to be asserted pretty much every time when
the exit code is asserted. This maybe could even be the same function if
its interface wouldn't get messy and unwieldly.