Added output assertions for tests in "applying"
For now the assertions are added mostly just to pass. A few most obvious
cases of wrong outputs were fixed (mostly when it was printing much more
than needed due to outputs of git commands called internally).
However, the output was generally left close to what it was before. It
will be rethought in the later commits as the point of this one is just
to establish existence of the necessary assertions in the tests.
(And this alone is a lot of work; so much that it will be spread onto
multiple commits.)

(Also a few errors in tests were fixed. The kind of errors where both
the exit code and the resulting repository state were fully correct, but
the output revealed that the intended code path didn't run at all.)