Fixed bugs caused by not quoting empty strings
In a few places, constructs like `${var+"$var"}` were not quoted.
At least one of those could cause a confirmed bug. (Fortunately, as far
as I can tell, the only effect of this bug was an error line in stderr.)

(No automatic tests for this were added, because it would require
checking the output of the script and this is already begin worked on
in a different branch.)