Added output assertions for tests in "creating"
This is the same situation as with the previous commit for "applying".
For now the assertions are added mostly just to pass. A few most obvious
cases of wrong outputs were fixed (mostly when it was printing much more
than needed due to outputs of git commands called internally).
However, the output was generally left close to what it was before. It
will be rethought in the later commits as the point of this one is just
to establish existence of the necessary assertions in the tests.
(And this alone is a lot of work; so much that it is being spread onto
multiple commits.)