Added output assertions for tests in "main_script"
This is the same situation as with the previous "intent-to-add" commit.
For now the assertions are added mostly just to pass. A few most obvious
cases of wrong outputs were fixed.
However, the output was generally left close to what it was before. It
will be rethought in the later commits as the point of this one is just
to establish existence of the necessary assertions in the tests.
(And this alone is a lot of work; so much that it is being spread onto
multiple commits.)