Added filter to tests for only changed
Added options `--altered` (`-a`) and `--since` (`-A`) to test scripts
`./tests/list.sh` and `./tests/run.sh`. (`--changed` would maybe be a
better name but it collided with other options.)
This reduces the need for manual filtering for specific tests when you
want to run only the ones that were changed recently.