Fixed the function that checks istash data files
The function decides which operation (`apply` or `pop`) is running right
now or whether the set of data files is incorrect. Before this fix, it
didn't take into account the data file that stores the commit with state
of the working copy, that was added to the scripts most recently.