Fixed an error where the list of files is too long
There was a problem with a `xargs` command that was passing the entire
input stream as a single argument to a command. The stream had no split
points so it was possible to exceed the maximum argument length.

Test for this runs rather long, so it is skipped on meticulousness <= 1
and it has only a few parametrizations.
(The test is quite unique because it's dynamic. It first asses the limit
of `xargs` and then generates enough files to always trigger it.)